Added comments, deleted function
I've commented the code in FR and EN.

I also tried to fix the issue with the groups :
- If the function check_group_if_layers_in was active : The groups weren't checked if a child was and the layers unchecked and then rechecked was displaying checked but hidden (???)
- If the function is not here, this does not change anything for the groups but that fixed the issue with hidden layers.

I don't understand why QGIS is activating the group i you check a layer in automatically in the layer tree but not thanks to the plugin. And as the groups are not documented in QGIS API i can't use it (or i didn't found it).

For now, the groups aren't supported. I would have liked to force every groups to be displayed even if any layers in were shown (that would have change nothing for the user) but I didn't successed.

Here's the challenge.